+++ src/resources/mod.rs
@@ -1,7 +1,6 @@
 pub use self::resource::*;
-pub use self::resource_view::*;
-pub use self::storage::*;
+
+pub(crate) use self::storage::*;
 
 mod resource;
-mod resource_view;
 mod storage;