+++ src/storage/mod.rs
@@ -1,12 +1,10 @@
-pub use self::change_ticks::*;
 pub use self::component_storage::*;
 pub use self::entity::*;
-pub use self::entity_sparse_array::*;
+pub use self::sparse_array::*;
 
 pub(crate) use self::entity_storage::*;
 
-mod change_ticks;
 mod component_storage;
 mod entity;
-mod entity_sparse_array;
 mod entity_storage;
+mod sparse_array;