+++ src/world/borrow.rs
@@ -1,7 +1,6 @@
 use crate::components::Component;
-use crate::query::ComponentView;
-use crate::resources::{Resource, ResourceCell, ResourceView};
-use crate::storage::{ComponentStorage, Ticks};
+use crate::resources::Resource;
+use crate::storage::ComponentStorage;
 use crate::utils::{impl_generic_1_16, panic_missing_comp, panic_missing_res};
 use crate::world::World;
 use atomic_refcell::{AtomicRef, AtomicRefMut};
@@ -12,15 +11,15 @@ pub type Comp<'a, T> = ComponentView<'a, T, AtomicRef<'a, ComponentStorage>>;
 /// Exclusive view over a component storage.
 pub type CompMut<'a, T> = ComponentView<'a, T, AtomicRefMut<'a, ComponentStorage>>;
 /// Shared view over a resource.
-pub type Res<'a, T> = ResourceView<T, AtomicRef<'a, ResourceCell>>;
+pub type Res<'a, T> = AtomicRef<'a, T>;
 /// Exclusive view over a resource.
-pub type ResMut<'a, T> = ResourceView<T, AtomicRefMut<'a, ResourceCell>>;
+pub type ResMut<'a, T> = AtomicRefMut<'a, T>;
 
 /// Trait used to borrow component storages and resources from a `World`.
 pub trait BorrowWorld<'a> {
     type Item;
 
-    fn borrow(world: &'a World, change_tick: Ticks) -> Self::Item;
+    fn borrow(world: &'a World) -> Self::Item;
 }
 
 impl<'a, 'b, T> BorrowWorld<'a> for Comp<'b, T>
@@ -29,13 +28,13 @@ where
 {
     type Item = Comp<'a, T>;
 
-    fn borrow(world: &'a World, change_tick: Ticks) -> Self::Item {
+    fn borrow(world: &'a World) -> Self::Item {
         let (storage, info) = world
             .component_storages()
             .borrow_with_info(&TypeId::of::<T>())
             .unwrap_or_else(|| panic_missing_comp::<T>());
 
-        unsafe { Comp::new(storage, info, world.tick(), change_tick) }
+        unsafe { Comp::new(storage, info) }
     }
 }
 
@@ -45,13 +44,13 @@ where
 {
     type Item = CompMut<'a, T>;
 
-    fn borrow(world: &'a World, change_tick: Ticks) -> Self::Item {
+    fn borrow(world: &'a World) -> Self::Item {
         let (storage, info) = world
             .component_storages()
             .borrow_with_info_mut(&TypeId::of::<T>())
             .unwrap_or_else(|| panic_missing_comp::<T>());
 
-        unsafe { CompMut::new(storage, info, world.tick(), change_tick) }
+        unsafe { CompMut::new(storage, info) }
     }
 }
 
@@ -61,12 +60,8 @@ where
 {
     type Item = Res<'a, T>;
 
-    fn borrow(world: &'a World, change_tick: Ticks) -> Self::Item {
-        world
-            .resource_storage()
-            .borrow(&TypeId::of::<T>())
-            .map(|cell| unsafe { Res::new(cell, world.tick(), change_tick) })
-            .unwrap_or_else(|| panic_missing_res::<T>())
+    fn borrow(world: &'a World) -> Self::Item {
+        world.resource_storage().borrow::<T>().unwrap_or_else(|| panic_missing_res::<T>())
     }
 }
 
@@ -76,12 +71,8 @@ where
 {
     type Item = ResMut<'a, T>;
 
-    fn borrow(world: &'a World, change_tick: Ticks) -> Self::Item {
-        world
-            .resource_storage()
-            .borrow_mut(&TypeId::of::<T>())
-            .map(|cell| unsafe { ResMut::new(cell, world.tick(), change_tick) })
-            .unwrap_or_else(|| panic_missing_res::<T>())
+    fn borrow(world: &'a World) -> Self::Item {
+        world.resource_storage().borrow_mut::<T>().unwrap_or_else(|| panic_missing_res::<T>())
     }
 }
 
@@ -91,11 +82,8 @@ where
 {
     type Item = Option<Res<'a, T>>;
 
-    fn borrow(world: &'a World, change_tick: Ticks) -> Self::Item {
-        world
-            .resource_storage()
-            .borrow(&TypeId::of::<T>())
-            .map(|cell| unsafe { Res::new(cell, world.tick(), change_tick) })
+    fn borrow(world: &'a World) -> Self::Item {
+        world.resource_storage().borrow::<T>()
     }
 }
 
@@ -105,11 +93,8 @@ where
 {
     type Item = Option<ResMut<'a, T>>;
 
-    fn borrow(world: &'a World, change_tick: Ticks) -> Self::Item {
-        world
-            .resource_storage()
-            .borrow_mut(&TypeId::of::<T>())
-            .map(|cell| unsafe { ResMut::new(cell, world.tick(), change_tick) })
+    fn borrow(world: &'a World) -> Self::Item {
+        world.resource_storage().borrow_mut::<T>()
     }
 }
 
@@ -121,8 +106,8 @@ macro_rules! impl_borrow_world {
         {
             type Item = ($($ty::Item,)+);
 
-            fn borrow(world: &'a World, change_tick: Ticks) -> Self::Item {
-                ($($ty::borrow(world, change_tick),)+)
+            fn borrow(world: &'a World) -> Self::Item {
+                ($($ty::borrow(world),)+)
             }
         }
     };